requirements.txt
neurosity
python-dotenv
numpy
pandas
